+++ lib/flutter_sliding_toast.dart
@@ -1,7 +1,9 @@
 library flutter_sliding_toast;
 
 export 'package:flutter_sliding_toast/src/sliding_toast.dart';
-export 'package:flutter_sliding_toast/src/toast_setting.dart';
-export 'package:flutter_sliding_toast/src/toast_style.dart';
-export 'package:flutter_sliding_toast/src/toast_controller.dart';
+export 'package:flutter_sliding_toast/src/popping_toast.dart';
+export 'package:flutter_sliding_toast/src/models/sliding_toast_setting.dart';
+export 'package:flutter_sliding_toast/src/models/popping_toast_setting.dart';
+export 'package:flutter_sliding_toast/src/models/toast_style.dart';
+export 'package:flutter_sliding_toast/src/models/toast_controller.dart';
 export 'package:flutter_sliding_toast/src/enums/toast_position.dart';