+++ lib/src/models/popping_toast_setting.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 
-import 'enums/toast_position.dart';
+import '../enums/toast_position.dart';
 
 /// The setting for the sliding animation
 @immutable
-class ToastSetting {
+class PoppingToastSetting {
   /// The maximum width of the toast
   /// Default is [80% of the screen width]
   final double? maxWidth;
@@ -13,12 +13,24 @@ class ToastSetting {
   /// Default is [40% of the screen height]
   final double? maxHeight;
 
-  /// The sliding behavior of the animation.
-  /// Default is [Curves.elasticOut]
-  final Curve curve;
+  /// Should the toast be faded or not
+  /// Default is true
+  final bool showFading;
 
-  /// How long the toast will slide.
-  /// Default is [Duration(seconds: 1)]
+  /// Should the toast be faded or not
+  /// Default is true
+  final bool showScaling;
+
+  /// The fading behavior of the toast.
+  /// Default is [Curves.easeOut]
+  final Curve fadingCurve;
+
+  /// The scaling behavior of the toast.
+  /// Default is [Curves.bounceOut]
+  final Curve scalingCurve;
+
+  /// How long the toast will fade.
+  /// Default is [Duration(milliSeconds: 500)]
   final Duration animationDuration;
 
   /// How long the toast will be displayed
@@ -30,22 +42,9 @@ class ToastSetting {
   /// Default is [true]
   final bool showReverseAnimation;
 
-  /// The height of the progress bar.
-  /// Default is [3]
-  /// ProgressBar height constraint is (max: 8, min:2)
-  final double progressBarHeight;
-
-  /// Whether to show the progress bar or not.
-  /// Default is [true]
-  final bool showProgressBar;
-
   /// Disable Multiple tap on the toast
   /// Default value is false
   final bool disableMultiTapping;
-
-  /// The position of the toast in the screen where it will slide from.
-  /// Default is [ToastPosition.bottom]
-  final ToastPosition toastStartPosition;
 
   /// The toast alignment in the screen where it will stop.
   /// If the [ToastPosition] is top,
@@ -60,40 +59,46 @@ class ToastSetting {
   /// Default is [EdgeInsets.all(10)]
   final EdgeInsets padding;
 
-  const ToastSetting({
+  const PoppingToastSetting({
     this.maxWidth,
     this.maxHeight,
-    this.curve = Curves.elasticOut,
-    this.animationDuration = const Duration(seconds: 1),
+    this.showFading = true,
+    this.showScaling = true,
+    this.fadingCurve = Curves.easeOut,
+    this.scalingCurve = Curves.elasticOut,
+    this.animationDuration = const Duration(microseconds: 500),
     this.displayDuration = const Duration(seconds: 1),
     this.showReverseAnimation = true,
-    this.showProgressBar = true,
-    this.progressBarHeight = 3,
     this.disableMultiTapping = false,
-    this.toastStartPosition = ToastPosition.bottom,
     this.toastAlignment = Alignment.bottomCenter,
     this.padding = const EdgeInsets.all(10),
   });
 
-  ToastSetting copyWith({
+  PoppingToastSetting copyWith({
     double? maxWidth,
     double? maxHeight,
-    Curve? curve,
+    bool? showFading,
+    bool? showScaling,
+    Curve? fadingCurve,
+    Curve? scalingCurve,
     Duration? animationDuration,
     Duration? displayDuration,
     bool? showReverseAnimation,
-    ToastPosition? toastStartPosition,
+    bool? disableMultiTapping,
     Alignment? toastAlignment,
     EdgeInsets? padding,
   }) {
-    return ToastSetting(
+    return PoppingToastSetting(
       maxWidth: maxWidth ?? this.maxWidth,
       maxHeight: maxHeight ?? this.maxHeight,
-      curve: curve ?? this.curve,
+      showFading: showFading ?? this.showFading,
+      showScaling: showScaling ?? this.showScaling,
+      fadingCurve: fadingCurve ?? this.fadingCurve,
+      scalingCurve: scalingCurve ?? this.scalingCurve,
       animationDuration: animationDuration ?? this.animationDuration,
       displayDuration: displayDuration ?? this.displayDuration,
       showReverseAnimation: showReverseAnimation ?? this.showReverseAnimation,
-      toastStartPosition: toastStartPosition ?? this.toastStartPosition,
+      disableMultiTapping: disableMultiTapping ?? this.disableMultiTapping,
       toastAlignment: toastAlignment ?? this.toastAlignment,
       padding: padding ?? this.padding,
     );